+++ models/response_type.go
@@ -15,7 +15,7 @@ type BlockListHeaderResponse struct {
 	Limit     int                 `json:"limit"`
 	BlockInfo *BlockRet           `json:"block_info,omitempty"`
 	List      []BlockListResponse `json:"list"`
-	Digits    int64               `json:"digits"`
+	Digits    int                 `json:"digits"`
 }
 
 type BlockListResponse struct {
@@ -213,7 +213,7 @@ type EcosystemDetailInfoResponse struct {
 	RegistrationType string            `json:"registration_type"`
 	WebPage          string            `json:"web_page"`
 	Social           map[string]string `json:"social"`
-	Digits           int64             `json:"digits"`
+	Digits           int               `json:"digits"`
 	BridgeInfo       *bridgeInfo       `json:"bridge_info"`
 }
 
@@ -245,7 +245,7 @@ type EcosystemTokenSymbolList struct {
 	Amount         decimal.Decimal `json:"amount"`
 	AccountedFor   decimal.Decimal `json:"accounted_for"`
 	TokenSymbol    string          `json:"token_symbol"`
-	Digits         int64           `json:"digits"`
+	Digits         int             `json:"digits"`
 	FrontCommittee bool            `json:"front_committee"`
 	Committee      bool            `json:"committee"`
 	Activation     bool            `json:"activation"`
@@ -328,7 +328,7 @@ type accountRatio struct {
 // EcoSystem Chart response
 type EcoTopTenHasTokenResponse struct {
 	TokenSymbol string         `json:"token_symbol"`
-	Digits      int64          `json:"digits"`
+	Digits      int            `json:"digits"`
 	Name        string         `json:"name"` //ecosystem name
 	List        []accountRatio `json:"list"`
 }
@@ -348,7 +348,7 @@ type EcoCirculationsResponse struct {
 	Name             string `json:"name"`
 	SupplyToken      string `json:"supply_token"`
 	Emission         string `json:"emission"`
-	Digits           int64  `json:"digits"`
+	Digits           int    `json:"digits"`
 
 	Change EcoCirculationsChangeResponse `json:"change"`
 }
@@ -370,7 +370,7 @@ type EcoGasFeeResponse struct {
 	Combustion  string `json:"combustion"`
 	TokenSymbol string `json:"token_symbol"`
 	Name        string `json:"name"`
-	Digits      int64  `json:"digits"`
+	Digits      int    `json:"digits"`
 }
 
 type EcoGasFeeChangeResponse struct {
@@ -379,7 +379,7 @@ type EcoGasFeeChangeResponse struct {
 	Time        []string `json:"time"`
 	GasFee      []string `json:"gas_fee"`
 	Combustion  []string `json:"combustion"`
-	Digits      int64    `json:"digits"`
+	Digits      int      `json:"digits"`
 }
 
 type EcoTxAmountDiffResponse struct {
@@ -387,12 +387,12 @@ type EcoTxAmountDiffResponse struct {
 	Name        string   `json:"name"` //ecosystem name
 	Time        []int64  `json:"time"`
 	Amount      []string `json:"amount"`
-	Digits      int64    `json:"digits"`
+	Digits      int      `json:"digits"`
 }
 
 type EcoTxGasFeeDiffResponse struct {
 	TokenSymbol    string   `json:"token_symbol"`
-	Digits         int64    `json:"digits"`
+	Digits         int      `json:"digits"`
 	Name           string   `json:"name"` //ecosystem name
 	Time           []int64  `json:"time"`
 	EcoGasAmount   []string `json:"eco_gas_amount"`
@@ -408,14 +408,14 @@ type AccountTotalAmountChart struct {
 	LockAmount       decimal.Decimal `json:"lock_amount"`
 	LockRatio        float64         `json:"lock_ratio"`
 	TokenSymbol      string          `json:"token_symbol"`
-	Digits           int64           `json:"digits"`
+	Digits           int             `json:"digits"`
 }
 
 type AccountAmountChangePieChart struct {
 	Outcome     decimal.Decimal `json:"outcome" gorm:"column:outcome"`
 	Income      decimal.Decimal `json:"income" gorm:"column:income"`
 	TokenSymbol string          `json:"token_symbol" gorm:"column:token_symbol"`
-	Digits      int64           `json:"digits"`
+	Digits      int             `json:"digits"`
 }
 
 type AccountAmountChangeBarChart struct {
@@ -425,7 +425,7 @@ type AccountAmountChangeBarChart struct {
 	Outcome     []string `json:"outcome"`
 	Income      []string `json:"income"`
 	Balance     []string `json:"balance"`
-	Digits      int64    `json:"digits"`
+	Digits      int      `json:"digits"`
 }
 
 type AccountTxChart struct {
@@ -686,7 +686,7 @@ type NodeVoteResponse struct {
 type GasFee struct {
 	Amount      string `json:"amount"`
 	TokenSymbol string `json:"token_symbol"`
-	Digits      int64  `json:"digits"`
+	Digits      int    `json:"digits"`
 }
 
 type NodeBlockListResponse struct {
@@ -820,7 +820,7 @@ type utxoDetail struct {
 	Amount      string `json:"amount"`
 	Hash        string `json:"hash"`
 	TokenSymbol string `json:"token_symbol"`
-	Digits      int64  `json:"digits"`
+	Digits      int    `json:"digits"`
 }
 
 type FuelRateResponse struct {
@@ -833,7 +833,7 @@ type FeeInfo struct {
 	Recipient   string `json:"recipient"`
 	Amount      string `json:"amount"`
 	TokenSymbol string `json:"tokenSymbol"`
-	Digits      int64  `json:"digits"`
+	Digits      int    `json:"digits"`
 }
 
 type combusionInfo struct {
@@ -842,7 +842,7 @@ type combusionInfo struct {
 	Amount      decimal.Decimal `json:"amount"`
 	TokenSymbol string          `json:"tokenSymbol"`
 	Rate        string          `json:"rate"`
-	Digits      int64           `json:"digits"`
+	Digits      int             `json:"digits"`
 }
 
 type FeesInfo struct {
@@ -852,7 +852,7 @@ type FeesInfo struct {
 	Fees        FeeInfo          `json:"fees"`
 	Taxes       FeeInfo          `json:"taxes"`
 	Combustion  combusionInfo    `json:"combustion"`
-	Digits      int64            `json:"digits"`
+	Digits      int              `json:"digits"`
 }
 
 type UtxoExplorerInfo struct {
@@ -881,7 +881,7 @@ type UtxoExplorer struct {
 	BasisGasFee FeesInfo     `json:"basis_gas_fee"`
 	EcoGasFee   FeesInfo     `json:"eco_gas_fee"`
 	Size        string       `json:"size"`
-	Digits      int64        `json:"digits"`
+	Digits      int          `json:"digits"`
 }
 
 type AccountTxListResponse struct {
